pyproject.toml
[project]
name = "subsentry"
version = "0.1.0"
description = "Local-first, privacy-first personal finance analyzer with optional local LLM + MCP server."
requires-python = ">=3.10"
readme = "README.md"
authors = [{name="SubSentry"}]
dependencies = [
  "streamlit>=1.36",
  "pydantic>=2.6",
  "sqlalchemy>=2.0",
  "python-dateutil>=2.9",
  "rapidfuzz>=3.6",
  "cryptography>=42.0",
  "argon2-cffi>=23.1",
  "pandas>=2.2",
  "numpy>=1.26",
  "mcp>=1.7.0",
  "requests>=2.32",
  "plotly>=5.20",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6",
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "mypy>=1.10",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
